Prebid 9: rename GDPR modules (#11521)
