+++ test/spec/modules/BTBidAdapter_spec.js
@@ -7,10 +7,10 @@ import 'modules/currency.js';
 import 'modules/userId/index.js';
 import 'modules/multibid/index.js';
 import 'modules/priceFloors.js';
-import 'modules/consentManagement.js';
+import 'modules/consentManagementTcf.js';
 import 'modules/consentManagementUsp.js';
 import 'modules/consentManagementGpp.js';
-import 'modules/gdprEnforcement.js';
+import 'modules/tcfControl.js';
 import 'modules/gppControl_usnat.js';
 import 'modules/schain.js';
