+++ test/spec/modules/consentManagement_spec.js
@@ -8,7 +8,7 @@ import {
   setConsentConfig,
   staticConsentData,
   userCMP
-} from 'modules/consentManagement.js';
+} from 'modules/consentManagementTcf.js';
 import {gdprDataHandler} from 'src/adapterManager.js';
 import * as utils from 'src/utils.js';
 import {config} from 'src/config.js';
