+++ test/spec/modules/conversantBidAdapter_spec.js
@@ -8,7 +8,7 @@ import 'src/prebid.js'
 import 'modules/currency.js';
 import 'modules/userId/index.js'; // handles eids
 import 'modules/priceFloors.js';
-import 'modules/consentManagement.js';
+import 'modules/consentManagementTcf.js';
 import 'modules/consentManagementUsp.js';
 import 'modules/schain.js'; // handles schain
 import {hook} from '../../../src/hook.js'
